[Build 616] [Update] Test du son : ok

+++ Assets/Script/FrameWork/Song.cs
@@ -30,7 +30,7 @@ public class Song {
 	public int numberOfJumps;
 	public int numberOfHands;
 	public double duration;
-	public float[] intensityGraph;
+	public double[] intensityGraph;
 	
 	public double stepPerSecondAverage;
 	public double stepPerSecondMaximum;
